update user repository methods to require user id

+++ app/linkservice/delete-link.go
@@ -1,5 +1,10 @@
 package linkservice
 
-func (s *Service) DeleteLink(id uint) (uint, error) {
+import "github.com/andyinabox/linkydink/app"
+
+func (s *Service) DeleteLink(userId uint, id uint) (uint, error) {
+	if userId == 0 {
+		return 0, app.ErrMissingUserId
+	}
 	return s.lr.DeleteLink(id)
 }
